Move styled components to styled
